also raise exception for errno 19
Errno 19 happens on my Linux system when pulling the cable. Its defined
as `ENODEV`="No such device" which makes sense for my situation.
This situation should not be silently skipped. Plus: I got an exception
a few lines later when trying to use `length` which is undefined because
of the try-block being skipped due to the exception.